start and stop window
