+++ packages/client/src/containers/GameContainer/index.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useEffect, useCallback } from "react";
+import React, { useRef, useEffect, useCallback, forwardRef, useImperativeHandle } from "react";
 import type { Dispatch, SetStateAction } from "react";
 
 import Game from "../../core/game/Game";
@@ -9,9 +9,13 @@ const debug = debugResolve("GameContainer");
 type TGameProps = {
     height: string;
     width: string;
-    newGame: boolean;
     setScore: Dispatch<SetStateAction<number>>;
-    setIsGameOver: Dispatch<SetStateAction<boolean>>;
+    onGameOver: () => void;
+};
+
+export type GameForwardProps = {
+    start: () => void;
+    restart: () => void;
 };
 
 const KEY_CODE_ARROW_UP = "ArrowUp";
@@ -23,77 +27,76 @@ const KEY_CODE_S = "KeyS";
 const KEY_CODE_A = "KeyA";
 const KEY_CODE_D = "KeyD";
 
-const GameContainer: React.FC<TGameProps> = ({
-    height,
-    width,
-    newGame,
-    setScore,
-    setIsGameOver
-}) => {
-    const canvasRef = useRef(null);
-    const gameRef = useRef<Game | null>(null);
+const GameContainer = forwardRef<GameForwardProps, TGameProps>(
+    ({ height, width, setScore, onGameOver }, ref) => {
+        const canvasRef = useRef(null);
+        const gameRef = useRef<Game | null>(null);
 
-    const handleKeyPress = useCallback(
-        (event: KeyboardEvent) => {
-            const { code } = event;
+        useImperativeHandle(ref, () => {
+            return {
+                restart: () => {
+                    if (canvasRef.current && gameRef.current) {
+                        gameRef.current.restart();
+                    }
+                },
+                start: () => {
+                    if (canvasRef.current) {
+                        gameRef.current = new Game(canvasRef.current);
+                    }
 
-            debug("code", code);
-
-            if (gameRef.current) {
-                switch (code) {
-                    case KEY_CODE_ARROW_UP:
-                    case KEY_CODE_W:
-                        gameRef.current.moveUp();
-                        break;
-                    case KEY_CODE_ARROW_DOWN:
-                    case KEY_CODE_S:
-                        gameRef.current.moveDown();
-                        break;
-                    case KEY_CODE_ARROW_LEFT:
-                    case KEY_CODE_A:
-                        gameRef.current.moveLeft();
-                        break;
-                    case KEY_CODE_ARROW_RIGHT:
-                    case KEY_CODE_D:
-                        gameRef.current.moveRight();
-                        break;
+                    gameRef.current?.start();
                 }
+            };
+        });
 
-                setScore(gameRef.current.updateScore());
+        const handleKeyPress = useCallback(
+            (event: KeyboardEvent) => {
+                const { code } = event;
 
-                const isGameOver = gameRef.current.checkEndConditions();
+                debug("code", code);
 
-                if (isGameOver) {
-                    setIsGameOver(true);
+                if (gameRef.current) {
+                    switch (code) {
+                        case KEY_CODE_ARROW_UP:
+                        case KEY_CODE_W:
+                            gameRef.current.moveUp();
+                            break;
+                        case KEY_CODE_ARROW_DOWN:
+                        case KEY_CODE_S:
+                            gameRef.current.moveDown();
+                            break;
+                        case KEY_CODE_ARROW_LEFT:
+                        case KEY_CODE_A:
+                            gameRef.current.moveLeft();
+                            break;
+                        case KEY_CODE_ARROW_RIGHT:
+                        case KEY_CODE_D:
+                            gameRef.current.moveRight();
+                            break;
+                    }
+
+                    setScore(gameRef.current.updateScore());
+
+                    const isGameOver = gameRef.current.checkEndConditions();
+
+                    if (isGameOver) {
+                        onGameOver();
+                    }
                 }
-            }
-        },
-        [setScore]
-    );
+            },
+            [setScore]
+        );
 
-    useEffect(() => {
-        if (canvasRef.current) {
-            gameRef.current = new Game(canvasRef.current);
-        }
+        useEffect(() => {
+            document.addEventListener("keydown", handleKeyPress);
 
-        gameRef.current?.start();
-    }, []);
+            return () => {
+                document.removeEventListener("keydown", handleKeyPress);
+            };
+        }, [handleKeyPress]);
 
-    useEffect(() => {
-        if (newGame && canvasRef.current && gameRef.current) {
-            gameRef.current.restart();
-        }
-    }, [newGame]);
-
-    useEffect(() => {
-        document.addEventListener("keydown", handleKeyPress);
-
-        return () => {
-            document.removeEventListener("keydown", handleKeyPress);
-        };
-    }, [handleKeyPress]);
-
-    return <canvas width={width} height={height} ref={canvasRef}></canvas>;
-};
+        return <canvas width={width} height={height} ref={canvasRef}></canvas>;
+    }
+);
 
 export default GameContainer;
