--- packages/client/src/pages/Game/components/ButtonsBoard/styles.scss
@@ -1,24 +1,0 @@
-.buttonsBoard {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 10px;
-    padding: 13px;
-
-    &__button {
-        background-color: $purple-dark;
-        color: $white-primary;
-        border-radius: 3px;
-        line-height: 42px;
-        font-weight: bold;
-        font-size: 18px;
-        padding: 0 20px;
-        width: 100%;
-        text-decoration: none;
-        cursor: pointer;
-        border: none;
-
-        &:active {
-            opacity: 0.7;
-        }
-    }
-}
