--- packages/client/src/pages/Game/components/Main/styles.scss
@@ -1,5 +1,0 @@
-.main {
-    min-width: 500px;
-    min-height: 500px;
-    border: 1px solid orange;
-}
