+++ packages/client/src/pages/Game/styles.scss
@@ -1,4 +1,5 @@
 .game {
+    position: relative;
     display: grid;
     grid-template-rows: repeat(3, max-content);
     gap: 10px;
@@ -7,5 +8,50 @@
     &:fullscreen {
         background-color: $yellow-color;
         padding: 2rem;
+    }
+}
+
+.game-filter {
+    filter: blur(4px);
+    &--off {
+        filter: unset;
+    }
+}
+
+.game-over {
+    max-width: 400px;
+    position: absolute;
+    left: 50px;
+    top: 240px;
+    font-weight: 600;
+    font-size: 1.5rem;
+    color: $purple-dark;
+}
+
+.buttonsBoard {
+    &__button {
+        z-index: 10;
+        background-color: $purple-dark;
+        color: $white-primary;
+        border-radius: 3px;
+        line-height: 42px;
+        font-weight: bold;
+        font-size: 18px;
+        padding: 0 20px;
+        width: 100%;
+        text-decoration: none;
+        cursor: pointer;
+        border: none;
+
+        &:active {
+            opacity: 0.7;
+        }
+
+        &--start {
+            max-width: 300px;
+            position: absolute;
+            left: 50px;
+            top: 280px;
+        }
     }
 }
