--- packages/client/src/pages/end/end.tsx
@@ -1,6 +1,0 @@
-import "./styles.scss";
-import { EndScreen } from "../../containers/EndScreen";
-
-export const EndPage = () => {
-    return <EndScreen />;
-};
