--- packages/client/src/pages/end/index.ts
@@ -1,1 +1,0 @@
-export * from "./end";
