--- packages/client/src/pages/start/index.ts
@@ -1,1 +1,0 @@
-export * from "./start";
