--- packages/client/src/pages/start/start.tsx
@@ -1,8 +1,0 @@
-import { withProtection } from "@containers";
-
-import { StartScreen } from "../../containers/StartScreen";
-import "./styles.scss";
-
-export const StartPage = withProtection(() => {
-    return <StartScreen />;
-});
