--- packages/client/src/pages/start/styles.scss
@@ -1,45 +1,0 @@
-.start__wrapper {
-    box-shadow: 0 0 10px 2px rgba(0, 0, 0, 0.75);
-    min-width: 20rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 20px 0;
-
-    .logo__wrapper {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin-bottom: 20px;
-
-        .logo {
-            display: flex;
-            flex-wrap: wrap;
-            width: 55px;
-            height: 55px;
-            column-gap: 5px;
-            row-gap: 5px;
-
-            p {
-                margin: 0;
-                width: 25px;
-                height: 25px;
-                border-radius: 25px;
-                text-align: center;
-
-                &:nth-child(1),
-                &:nth-child(4) {
-                    background: $black-primary;
-                    color: $white-primary;
-                }
-
-                &:nth-child(2),
-                &:nth-child(3) {
-                    background: $blue-color;
-                    color: $white-primary;
-                }
-            }
-        }
-    }
-}
